transportation-maritime/maritime_solution/ai_services/sensor_simulator/analytics/src/requirements.txt
shapely==2.0.1
requests==2.28.2
influxdb==5.3.1
eclipse-zenoh==0.10.0rc0
pyyaml==6.0.1
fire==0.5.0
asyncio==3.4.3
asyncua==1.0.6
numpy==1.25.2
pandas==1.5.3
pycaret==3.2.0